refactor: segment storage owner consumers by domain
