+++ rustfs/src/storage/head_prefix.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use super::ECStore;
-use crate::storage::contract::list::ListOperations as _;
+use crate::storage::storage_api::head_prefix_consumer::contract::list::ListOperations as _;
 use std::sync::Arc;
 
 /// Determines if the key "looks like a prefix" (ends with `/`).
