+++ rustfs/src/storage/s3_api/bucket.rs
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::storage::contract::{
+use crate::storage::s3_api::common::rustfs_owner;
+use crate::storage::storage_api::s3_api_consumer::bucket::contract::{
     bucket::BucketInfo,
     list::{ListObjectVersionsInfo as StorageListObjectVersionsInfo, ListObjectsV2Info as StorageListObjectsV2Info},
 };
-use crate::storage::s3_api::common::rustfs_owner;
-use crate::storage::to_s3s_etag;
+use crate::storage::storage_api::s3_api_consumer::bucket::to_s3s_etag;
 use percent_encoding::percent_decode_str;
 use s3s::dto::{
     Bucket, CommonPrefix, DeleteMarkerEntry, EncodingType, ListBucketsOutput, ListObjectVersionsOutput, ListObjectsOutput,
@@ -27,7 +27,7 @@ use s3s::{S3Error, S3ErrorCode};
 use tracing::debug;
 use urlencoding::encode;
 
-use crate::storage::StorageObjectInfo as ObjectInfo;
+use crate::storage::storage_api::s3_api_consumer::bucket::StorageObjectInfo as ObjectInfo;
 
 const S3_MAX_KEYS: i32 = 1000;
 
@@ -395,9 +395,9 @@ mod tests {
         build_list_object_versions_output, build_list_objects_output, build_list_objects_v2_output,
         parse_list_object_versions_params, parse_list_objects_v2_params,
     };
-    use crate::storage::StorageObjectInfo as ObjectInfo;
-    use crate::storage::contract::bucket::BucketInfo;
     use crate::storage::s3_api::common::rustfs_owner;
+    use crate::storage::storage_api::s3_api_consumer::bucket::StorageObjectInfo as ObjectInfo;
+    use crate::storage::storage_api::s3_api_consumer::bucket::contract::bucket::BucketInfo;
     use s3s::S3ErrorCode;
     use s3s::dto::{CommonPrefix, EncodingType, ListObjectsV2Output, Object};
     use time::OffsetDateTime;
