+++ rustfs/src/storage/s3_api/multipart.rs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::storage::contract::multipart::{ListMultipartsInfo, ListPartsInfo};
 use crate::storage::s3_api::common::{rustfs_initiator, rustfs_owner};
-use crate::storage::to_s3s_etag;
+use crate::storage::storage_api::s3_api_consumer::multipart::contract::multipart::{ListMultipartsInfo, ListPartsInfo};
+use crate::storage::storage_api::s3_api_consumer::multipart::to_s3s_etag;
 use s3s::dto::{CommonPrefix, ListMultipartUploadsOutput, ListPartsOutput, MultipartUpload, Part, Timestamp};
 use s3s::{S3Error, S3ErrorCode};
 
@@ -195,9 +195,11 @@ mod tests {
         MAX_MULTIPART_UPLOADS_LIST, build_list_multipart_uploads_output, build_list_parts_output,
         parse_list_multipart_uploads_params, parse_list_parts_params, parse_upload_part_number,
     };
-    use crate::storage::contract::multipart::{ListMultipartsInfo, ListPartsInfo, MultipartInfo, PartInfo};
     use crate::storage::s3_api::common::{rustfs_initiator, rustfs_owner};
-    use crate::storage::to_s3s_etag;
+    use crate::storage::storage_api::s3_api_consumer::multipart::contract::multipart::{
+        ListMultipartsInfo, ListPartsInfo, MultipartInfo, PartInfo,
+    };
+    use crate::storage::storage_api::s3_api_consumer::multipart::to_s3s_etag;
     use s3s::S3ErrorCode;
     use s3s::dto::Timestamp;
     use time::OffsetDateTime;
